requirements.txt
Django==1.11
django-tagging==0.4.6
MySQL-python==1.2.5
pytz==2017.3
